Refactor to allow simple matchers
