+++ src/VCR/RequestMatchers/MethodMatcher.php
@@ -4,20 +4,16 @@ namespace VCR\RequestMatchers;
 
 use VCR\Request;
 
-class MethodMatcher extends RequestMatcher implements IRequestMatcher {
+class MethodMatcher extends SimpleMatcher implements IRequestMatcher {
     public function getName() {
         return "method";
     }
 
-    public function match(Request $first, Request $second) {
-        $equal = $first->getMethod() == $second->getMethod();
-        if ($this->getMatchObserver() && $this->getMatchObserver()->shouldObserve()) {
-            $this->getMatchObserver()->markMismatch($first, $second, $this);
-        }
-        return $equal;
+    public function getMismatchMessagePrefix() {
+        return "Method";
     }
 
-    public function getMismatchMessage(Request $first, Request $second) {
-        return $this->buildSimpleMismatchMessage('Method', $first->getMethod(), $second->getMethod());
+    public function getRequestValue(Request $request) {
+        return $request->getMethod();
     }
 }
