+++ src/VCR/RequestMatchers/UrlMatcher.php
@@ -4,20 +4,16 @@ namespace VCR\RequestMatchers;
 
 use VCR\Request;
 
-class UrlMatcher extends RequestMatcher implements IRequestMatcher {
+class UrlMatcher extends SimpleMatcher implements IRequestMatcher {
     public function getName() {
         return "url";
     }
 
-    public function match(Request $first, Request $second) {
-        $equal = $first->getUrl() == $second->getUrl();
-        if ($this->getMatchObserver() && $this->getMatchObserver()->shouldObserve()) {
-            $this->getMatchObserver()->markMismatch($first, $second, $this);
-        }
-        return $equal;
+    public function getMismatchMessagePrefix() {
+        return "URL";
     }
 
-    public function getMismatchMessage(Request $first, Request $second) {
-        return $this->buildSimpleMismatchMessage('URL', $first->getUrl(), $second->getUrl());
+    public function getRequestValue(Request $request) {
+        return $request->getUrl();
     }
 }
